+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Breaking Changes
 * [system contract] 系统合约重构成预编译合约形式
-* [other] 删除eip，DAO等版本升级的Hard Fork和兼容性检查;删除Rinkeby，Testnet;
+* [other] 删除eip，DAO等版本升级的Hard Fork和兼容性检查;
+* [other] 删除Rinkeby，Testnet;删除ChainConfig的EmptyBlock设置
 
 ### Improvements
 * [other] 版本管理采用mod模型 - 汤涌，于宗坤，杜满想